server/requirements.txt
Flask>=3.0,<4.0
pynput>=1.7,<2.0
pystray>=0.19,<1.0
Pillow>=10.0,<11.0
pyinstaller>=6.0,<7.0
